Revert to single types implementing Dialer + Listener.
Where appropriate, e.g. TCP or UDS, have a single Config which
implements both, Dialer and Listener, instead of using separate types.

This commit also removes the various combinators of `core::Transport`
and instead disambiguates them directly on `ListenerExt` and
`DialerExt` which is similar to what is done for upgrade combinators.