Just `Upgrade`.
Instead of `InboundUpgrade` and `OutboundUpgrade` this commit reinstates
a single connection upgrade trait called `Upgrade`. The rationale behind
this is that upgrades are mere (asynchronous) transformations and if
they are applied to inbound or outbound connections is not their
concern. The previous motivation for splitting the trait was the forced
unification of output types which may differ for inbound and outbound
upgrades, but rather than making upgrades directional (either by
splitting the trait, or by passing in an `Endpoint` parameter), they
just perform the upgrade and it is up to the user to apply them in a
correct way. If a protocol acts differently for inbound and outbound
connections and produces different outputs, they are written as two
different `Upgrade` impls and applied accordingly using `apply_inbound`
or `apply_outbound`. This change also restores a sane way to or together
multiple protocols, which was previously not sound.