Keep copy of dialing context in stream muxer.
Since `QuicConfig::dial` consumes the context we better keep a copy in
the future and subsequently the stream muxer to make sure it is not
dropped prematurely. For listening it is assumed that the listener
outlives the connections, hence we do not copy the context from a
listening stream into each stream muxer.