Return a sequence of `Multiaddr`s in `listen_on`.
This is part of the proposed changes to the `Transport` trait as
outlined in [1].

To allow transport implementations to listen on multiple addresses we
now report all those addresses as a sequence of multi-addresses instead
of supporting only a single one.

Instead of making `Transport::listen_on` return a generic iterator a
concrete type `MultiaddrSeq` is introduced to support non-emptiness.

[1]: https://github.com/libp2p/rust-libp2p/issues/794#issuecomment-475950601