Address review concerns.
1. Remove listener `Id`.
2. Bring back `ConnectedPoint` in `Transport::and_then` and
   `Transport::map`.
3. Safe polling in tests.

The listener ID was introduced because previously the listener address
was also used for identification purposes, e.g. in tests. Without being
able to refer to individual listeners, some tests made no longer sense
and had to be removed.

The replacement of `ConnectedPoint` by `Endpoint` was a side-effect of
some previous iteration of this PR which did not make the listen address
available in upgrades. This got eventually corrected, but I forgot to
bring back the `ConnectedPoint`.

Safe polling of a `RawSwarm` is a bit of a nuisance because it does not
implement `Stream`. I therefore wrapped the poll call in a future which
is being waited on. The polling of the swarm in some Kademlia tests is
also corrected using the regular `Stream` API.