Remove `SwarmEvent`.
As per review feedback the `SwarmEvent` is hereby removed. Instead tests
use randomly generated port numbers (collisions may occur especially
in the identify test which only uses the ephemeral TCP port range).

To me it seems the swarm API should be changed to make the dynamic set
of listen addresses accessible outside of `NetworkBehaviour`
implementations, be it for tests or for other use-cases. The sort of
port number sharing used now looks wrong to me.