Change `PingHandler::connection_keep_alive`.
Instead of `KeepAlive::No` we ask to keep the connection alive for at
least one second. The reason for this change is that
`tokio_timer::Delay` only guarantees millisecond precision and while we
try to make the `OutboundSubstreamRequest` as soon as possible, the
handler may in fact return `NotReady` before 1 millisecond has passed.
This in turn will make the `NodeHandler` close the connection
immediately and the pinging never even begins.