Add support for prefixed DHT keys of peer IDs.
Abstractly, there is a desire for node discovery that is targeted
at a specific subset of nodes in the DHT, whereby this subset is
associated with some identifier. That is, knowing such an identifier
for a subset of nodes, a particular node wants to discover other
peers in the DHT within this subset.

Concretely, in https://github.com/libp2p/rust-libp2p/issues/1087,
it is mentioned that a validator in Polkadot wants to discover
(at least some of) the nodes of a specific parachain it is (temporarily)
assigned to, and be able to do so in a targeted manner, i.e. without
randomly walking the DHT until it finds some. In that context, the
validator knows about an identifier for the parachain and the general
idea described in that issue seems to be around ways to tag or group
nodes of a parachain together in the DHT based on such an identifier.

To that end, this is an implementation that generally permits keys in
the Kademlia DHT to be assigned prefixes, corresponding to such an
identifier / namespace shared by multiple nodes and thus by multiple
DHT keys.