Make libp2p-quic more robust
There may be a rare, intermittent hang under certain conditions, but I
am having trouble reproducing it, and in libp2p it would be harmless ―
it would result in a timeout and immediate reconnection, which would
almost always succeed.  Libp2p needs to be able to handle such errors
anyway, since they can happen due to network problems, for example.
Therefore, I consider the futures portion of this code to be ready for
review.