Adapt to changed rustls API
‘ClientCertVerifier::client_auth_root_subjects()’ previously returned
a ‘Vec’, but now returns an ‘Option<Vec>’.  As a result,
‘Default::default()’ is now ‘None’, not ‘vec![]’.  This causes the
connection to be aborted instead of allowing any subject.

Fix this by explicitly returning ‘Some(vec![])’.

There is still a bug: if the TLS handshake fails, the code will panic
when it fails to obtain the nonexistent peer certificate.  This will be
fixed in the next commit.