core/src/connection: Bound task event buffer during conn establishment
This commit adds a size limit to the event buffer that is filled during
connection establishment and forwarded in bulk to the handler once the
connection is established.

When the size limit is reached an `EventBufferLimitReached` error is
triggered and send to the manager which will in turn drop its connection
to the task signaling it to shutdown.

The above buffer is already bound by the connection establishment
timeout. This commit adds an additional size bound to the buffer.