Use pure rust libraries to get host interfaces
For the sake of not having to compile any C code, let's try use only
pure Rust dependencies.  To get the host interface information we can
use the `pnet` and `ipnetwork` libraries.  These are both pure Rust.

`pnet` v0.25 depends on `ipnetwork` v0.15 so we use that also instead of
the latest `ipnetwork` v0.16.